storyscaffold/storysymbiosis/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
anthropic>=0.30.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
Pillow>=10.0.0
